server/requirements.txt
amqp==2.5.2
argh==0.26.2
astroid==2.4.1
autopep8==1.5.2
billiard==3.6.3.0
blinker==1.4
cachelib==0.1
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Session==0.3.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathtools==0.1.2
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.2
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.23.0
six==1.14.0
toml==0.10.1
urllib3==1.25.9
vine==1.3.0
watchdog==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
